changed package for Test's classes
