+++ src/test/java/app/dao/SaveInvitationDaoTest.java
@@ -1,7 +1,6 @@
-package app.dao.Invitation;
+package app.dao;
 
 import app.entities.Invitation;
-import app.exceptions.EntityNotFoundException;
 import org.dbunit.Assertion;
 import org.dbunit.dataset.DataSetException;
 import org.dbunit.dataset.IDataSet;
@@ -12,23 +11,23 @@ import org.junit.Test;
 import java.sql.Date;
 import java.sql.SQLException;
 
-public class UpdateInvitationDaoTest extends InitilizationInvitationDaoTest {
+public class SaveInvitationDaoTest extends InitilizationInvitationDaoTest {
 
     @Test
-    public void updateInvitation() {
-        invitationDao.edit(new Invitation(
-                1, 1, "email",
+    public void saveInvitation() {
+        invitationDao.save(new Invitation(
+                5, 1, "email@gmail.com",
                 "invatationCode", new Date(2019,3,3), "status"
         ));
         try {
             IDataSet iDataSet = new FlatXmlDataSetBuilder().build(getClass().getClassLoader()
-                    .getResourceAsStream(""));
+                    .getResourceAsStream("datasets/Invitation/save-dataset.xml"));
             ITable iTable = iDataSet.getTable(table);
 
-            IDataSet actualIDataSet = getMySqlConnection().createDataSet();
-            ITable actualITable = actualIDataSet.getTable(table);
+            IDataSet actualDataSet = getMySqlConnection().createDataSet();
+            ITable actualITAble = actualDataSet.getTable(table);
 
-            Assertion.assertEquals(iTable, actualITable);
+            Assertion.assertEquals(iTable, actualITAble);
         } catch (DataSetException e) {
             e.printStackTrace();
         } catch (SQLException e) {
@@ -36,13 +35,5 @@ public class UpdateInvitationDaoTest extends InitilizationInvitationDaoTest {
         } catch (Exception e) {
             e.printStackTrace();
         }
-    }
-
-    @Test(expected = EntityNotFoundException.class)
-    public void updateInvitationException() {
-        invitationDao.edit(new Invitation(
-                6, 1, "email",
-                "invatationCode", new Date(2019, 3, 3), "status"
-        ));
     }
 }
