+++ src/test/java/app/dao/UpdateInvitationDaoTest.java
@@ -1,6 +1,7 @@
-package app.dao.Invitation;
+package app.dao;
 
 import app.entities.Invitation;
+import app.exceptions.EntityNotFoundException;
 import org.dbunit.Assertion;
 import org.dbunit.dataset.DataSetException;
 import org.dbunit.dataset.IDataSet;
@@ -11,23 +12,23 @@ import org.junit.Test;
 import java.sql.Date;
 import java.sql.SQLException;
 
-public class SaveInvitationDaoTest extends InitilizationInvitationDaoTest {
+public class UpdateInvitationDaoTest extends InitilizationInvitationDaoTest {
 
     @Test
-    public void saveInvitation() {
-        invitationDao.save(new Invitation(
-                5, 1, "email@gmail.com",
+    public void updateInvitation() {
+        invitationDao.edit(new Invitation(
+                1, 1, "email",
                 "invatationCode", new Date(2019,3,3), "status"
         ));
         try {
             IDataSet iDataSet = new FlatXmlDataSetBuilder().build(getClass().getClassLoader()
-                    .getResourceAsStream("datasets/Invitation/save-dataset.xml"));
+                    .getResourceAsStream(""));
             ITable iTable = iDataSet.getTable(table);
 
-            IDataSet actualDataSet = getMySqlConnection().createDataSet();
-            ITable actualITAble = actualDataSet.getTable(table);
+            IDataSet actualIDataSet = getMySqlConnection().createDataSet();
+            ITable actualITable = actualIDataSet.getTable(table);
 
-            Assertion.assertEquals(iTable, actualITAble);
+            Assertion.assertEquals(iTable, actualITable);
         } catch (DataSetException e) {
             e.printStackTrace();
         } catch (SQLException e) {
@@ -35,5 +36,13 @@ public class SaveInvitationDaoTest extends InitilizationInvitationDaoTest {
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
+
+    @Test(expected = EntityNotFoundException.class)
+    public void updateInvitationException() {
+        invitationDao.edit(new Invitation(
+                6, 1, "email",
+                "invatationCode", new Date(2019, 3, 3), "status"
+        ));
     }
 }
